Refactor project structure: rename scripts, reorganize directories, and update file paths
