+++ skincarelib/datasets/.gitignore
@@ -1,2 +1,2 @@
-skincarelib/datasets/data/products_dataset_clean_tokens.csv
 data/products_dataset_clean_tokens.csv
+data/processed/products_dataset_clean_tokens.csv
