--- skincarelib/datasets/dummy.py
@@ -1,2 +1,0 @@
-def dummy_dataset():
-    return [1, 2, 3, 4]
