--- skincarelib/datasets/synonyms_v2.json
@@ -1,32 +1,0 @@
-{
-  "alchol denat": "alcohol denat",
-  "sd alcohol 40 b": "alcohol denat",
-  "sd alcohol 40-b": "alcohol denat",
-
-  "acsorbyl palmitate": "ascorbyl palmitate",
-  "ascorbyl glu-coside": "ascorbyl glucoside",
-  "ascorbyl tetraisopal-mitate": "ascorbyl tetraisopalmitate",
-
-  "caffine": "caffeine",
-
-  "camelia sinensis": "camellia sinensis",
-  "ca-mellia sinensis": "camellia sinensis",
-  "camillia sinensis": "camellia sinensis",
-
-  "cirtus aurantium dulcis (organic orange) fruit water": "citrus aurantium dulcis (organic orange) fruit water",
-  "citrus aurantuim dulcis (orange) fruit extract": "citrus aurantium dulcis (orange) fruit extract",
-  "citrus aurantium dulcis (orange)peel oil": "citrus aurantium dulcis (orange) peel oil",
-  "citrus limon (lemon)peel oil": "citrus limon (lemon) peel oil",
-  "citrus nobilis (mandarin orange) peeloil": "citrus nobilis (mandarin orange) peel oil",
-  "citrus bergamia (bergamot) esssential oil": "citrus bergamia (bergamot) essential oil",
-  "citrus auranitum bergamia (bergamot) fruit extract": "citrus aurantium bergamia (bergamot) fruit extract",
-
-  "butylene gylcol": "butylene glycol",
-  "dimethcone": "dimethicone",
-  "dimenthiconol": "dimethiconol",
-  "carmbomer": "carbomer",
-  "citricacid": "citric acid",
-  "benzyl acohol": "benzyl alcohol",
-  "benzyl alochol": "benzyl alcohol",
-  "amica montana flower extract": "arnica montana flower extract"
-}
